feat: generelize MenuProvider
AppMenuBridge now is app-dependent menu
