+++ src/ui/providers/MenuProvider.tsx
@@ -8,35 +8,6 @@ import React, {
 } from "react";
 import "../i18n";
 import { type Icon, type IconProps } from "@tabler/icons-react";
-import {
-    createAboutMenuItem,
-    createAboutSection,
-    createCheckForUpdatesMenuItem,
-    createCheckForUpdatesSection,
-    createCreateNewProjectMenuItem,
-    createExitMenuItem,
-    createExitSection,
-    createFileImportSection,
-    createFileRootMenuItem,
-    createGeneralHelpSection,
-    createHelpRootMenuItem,
-    createMolstarAppMenuItem,
-    createMolstarMenuItem,
-    createOnlyDevSection,
-    createOpenDevToolsMenuItem,
-    createOpenFileInViewerMenuItem,
-    createOpenRecentFileInViewerMenuItem,
-    createOpenUserDataFolderMenuItem,
-    createProjectActionsSection,
-    createReportIssueMenuItem,
-    createSettingsRootMenuItem,
-    createUtilitiesSection,
-} from "../config/systemMenuItems";
-import { useNavigate, type NavigateFunction } from "react-router-dom";
-import { useWorkspaceManagement } from "../features/workspace/hooks/useWorkspaceManagement";
-import { AboutDialogueContent } from "../features/about-dialogue/components/AboutDialogueContent";
-import { useDialogue, type ShowDialogueType } from "./DialogueProvider";
-import { pushErrorNotification } from "../services/NotificationService";
 
 /**
  * The action can be either:
@@ -155,32 +126,12 @@ export function useMenu() {
 
 type MenuProviderProps = {
     children: React.ReactNode;
+    initialMenu: Menu | (() => Menu);
 };
 
-export function MenuProvider({ children }: MenuProviderProps) {
-    // Use navigation.
-    const navigate = useNavigate();
-
-    // Use file management.
-    const {
-        loadFileAsStaging,
-        openFileExplorerAndLoadFileAsStaging,
-        createBlankFileAsStaging,
-    } = useWorkspaceManagement();
-
-    // Use dialogue.
-    const { showDialogue } = useDialogue();
-
+export function MenuProvider({ children, initialMenu }: MenuProviderProps) {
     // Menu state in the initial state.
-    const [menu, setMenu] = useState<Menu>(() =>
-        createInitialMenu(
-            navigate,
-            showDialogue,
-            openFileExplorerAndLoadFileAsStaging,
-            loadFileAsStaging,
-            createBlankFileAsStaging,
-        ),
-    );
+    const [menu, setMenu] = useState<Menu>(initialMenu);
 
     // For replace/restore functionality. Keeps original `MenuItem` to be restored.
     const backups = useRef<Map<string, MenuItem>>(new Map());
@@ -421,66 +372,4 @@ export function MenuProvider({ children }: MenuProviderProps) {
             {children}
         </MenuContext.Provider>
     );
-}
-
-function createInitialMenu(
-    navigate: NavigateFunction,
-    showDialogue: ShowDialogueType,
-    openFileExplorerAndLoadFileAsStaging: () => Promise<void>,
-    loadFileAsStaging: (fileData: Error | FileData[]) => Promise<void>,
-    createBlankFileAsStaging: () => void,
-): Menu {
-    return [
-        createFileRootMenuItem([
-            createProjectActionsSection([
-                createCreateNewProjectMenuItem(() => {
-                    createBlankFileAsStaging();
-                }),
-            ]),
-            createFileImportSection([
-                createOpenFileInViewerMenuItem(() =>
-                    openFileExplorerAndLoadFileAsStaging(),
-                ),
-                createOpenRecentFileInViewerMenuItem(async (path: string) => {
-                    const result = await window.electron.getFileData([path]);
-                    loadFileAsStaging(result);
-                }),
-            ]),
-            createUtilitiesSection([
-                createOpenUserDataFolderMenuItem(async () => {
-                    const result =
-                        await window.electron.requestToOpenUserDataFolder();
-                    if (result instanceof Error) {
-                        pushErrorNotification(
-                            `Not able to open user data folder! Details: <${result.message}>.`,
-                        );
-                    }
-                }),
-            ]),
-            createOnlyDevSection("file-dev", "For developers", [
-                createOpenDevToolsMenuItem(),
-            ]),
-            createExitSection([createExitMenuItem()]),
-        ]),
-        createSettingsRootMenuItem(navigate),
-        createHelpRootMenuItem([
-            createGeneralHelpSection([
-                createMolstarAppMenuItem(),
-                createMolstarMenuItem(),
-                createReportIssueMenuItem(),
-            ]),
-            createCheckForUpdatesSection([createCheckForUpdatesMenuItem()]),
-            createAboutSection([
-                createAboutMenuItem(async () => {
-                    await showDialogue({
-                        title: "About",
-                        showCloseButton: true,
-                        content: (close) => (
-                            <AboutDialogueContent close={close} />
-                        ),
-                    });
-                }),
-            ]),
-        ]),
-    ];
 }
